created 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,8 +9,8 @@
 int main(void)
 {
 	int i = 0;
-	int j = 1;
-	int k = 2;
+	long j = 1;
+	long k = 2;
 
 	while (i < 50)
 	{
